pyproject.toml
[project]
name = "polybot"
version = "0.1.0"
description = "Telegram bot for Polymarket trading"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "py-clob-client",
    "web3",
    "eth-account",
    "httpx",
    "aiosqlite",
    "cryptography",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
